+++ koin-projects/examples/android-samples/src/main/java/org/koin/sample/android/di/AppModule.kt
@@ -90,7 +90,7 @@ val javaModule = module {
         viewModel { CompatSimpleViewModel(get()) }
     }
     scope<JavaFragment> {
-
+        scoped { (id: String) -> ScopedPresenter(id, get()) }
     }
 }