+++ koin-projects/koin-core-ext/src/main/kotlin/org/koin/experimental/property/InjectProperties.kt
@@ -1,7 +1,7 @@
 package org.koin.experimental.property
 
 import org.koin.core.Koin
-import org.koin.core.context.KoinContextHandler
+import org.koin.core.context.GlobalContext
 import org.koin.core.scope.Scope
 import kotlin.reflect.KClass
 import kotlin.reflect.KMutableProperty0
@@ -10,7 +10,7 @@ fun <T : Any> T.inject(vararg properties: KMutableProperty0<*>) {
     properties.forEach {
         val prop = it as KMutableProperty0<Any>
         val type = prop.returnType.classifier as KClass<*>
-        val value = KoinContextHandler.get().get<Any>(type)
+        val value = GlobalContext.get().get<Any>(type)
         prop.set(value)
     }
 }