+++ koin-projects/koin-core/src/main/kotlin/org/koin/core/context/KoinContextHandler.kt
@@ -15,55 +15,8 @@
  */
 package org.koin.core.context
 
-import org.koin.core.Koin
-import org.koin.core.KoinApplication
-
 /**
  * Help hold any implementation of KoinContext
  */
-object KoinContextHandler {
-
-    private var _context: KoinContext? = null
-
-    private fun getContext(): KoinContext {
-        return _context ?: error("No Koin Context configured. Please use startKoin or koinApplication DSL. ")
-    }
-
-    /**
-     * Retrieve current KoinContext
-     */
-    fun get(): Koin = getContext().get()
-
-    /**
-     * Retrieve current KoinContext or null
-     */
-    fun getOrNull(): Koin? = _context?.getOrNull()
-
-    /**
-     * Register new KoinContext
-     *
-     * @throws IllegalStateException if already registered
-     */
-    fun register(koinContext: KoinContext) = synchronized(this) {
-        if (_context != null) {
-            error("A KoinContext is already started")
-        }
-        _context = koinContext
-    }
-
-    /**
-     * Start a Koin Application on current KoinContext
-     */
-    fun start(koinApplication: KoinApplication) {
-        getContext().setup(koinApplication)
-    }
-
-    /**
-     * Stop current KoinContext & clear it
-     */
-    fun stop() {
-        _context?.stop()
-        _context = null
-    }
-
-}
+@Deprecated("Better use GlobalContext")
+typealias KoinContextHandler = GlobalContext