+++ koin-projects/koin-core/src/test/java/org/koin/test/Asserts.kt
@@ -1,8 +1,8 @@
 package org.koin.test
 
 import org.junit.Assert.assertNull
-import org.koin.core.context.KoinContextHandler
+import org.koin.core.context.GlobalContext
 
 fun assertHasNoStandaloneInstance() {
-    assertNull(KoinContextHandler.getOrNull())
+    assertNull(GlobalContext.getOrNull())
 }