+++ koin-projects/koin-test/src/main/kotlin/org/koin/test/AutoCloseKoinTest.kt
@@ -23,7 +23,7 @@ import org.koin.core.context.stopKoin
  *
  * @author Arnaud Giuliani
  */
-abstract class AutoCloseKoinTest : KoinTest {
+interface AutoCloseKoinTest : KoinTest {
 
     @After
     fun autoClose() {